rename: move src/layouts to src/components/layouts
